src/cloud_functions/dev-microscope-classify/requirements.txt
c4v-py[gcloud,classification]==0.1.0.dev20221111
firebase-admin
google-cloud-bigquery
google-cloud-logging
pytz